update assets integration
