+++ store/wallet/types.ts
@@ -2,8 +2,10 @@ type Wallet = {
   address: string;
   name: string;
   symbol: string;
-  decimals: string;
+  decimals: number;
+  logo: string;
   balance: string;
+  balanceUSD: number;
   uiAmount: number;
   chainId: "base" | "solana";
 };
